requirements.txt
Django
dj-database-url
dj-static
django-toolbelt
gunicorn
psycopg2
static3
wsgiref
cython
tardis-sn
django-bootstrap3
django-graphos